refactor: split lecturers screen into hook and list components
